FIx code review issue.
