+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/conf/Config.java
@@ -166,17 +166,17 @@ public class Config {
              */
             public static boolean TRACE_DSL = false;
         }
-    }
+        public static class Customize {
+            /**
+             * Custom enhancement class configuration file path, recommended to use an absolute path.
+             */
+            public static String ENHANCE_FILE = "";
 
-    public static class Customize {
-        /**
-         * Custom enhancement class configuration file path, recommended to use an absolute path.
-         */
-        public static String ENHANCE_FILE = "";
-
-        /**
-         * Some information after custom enhancements, this configuration is used by the custom enhancement plugin.
-         */
-        public static Map<String, Object> CONTEXT = new HashMap<String, Object>();
+            /**
+             * Some information after custom enhancements, this configuration is used by the custom enhancement plugin.
+             * And using Map CONTEXT for avoiding classloader isolation issue.
+             */
+            public static Map<String, Object> CONTEXT = new HashMap<String, Object>();
+        }
     }
 }
