+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/plugin/PluginBootstrap.java
@@ -77,7 +77,7 @@ public class PluginBootstrap {
             }
         }
 
-        plugins.addAll(PluginSpiFactory.INSTANCE.load(AgentClassLoader.getDefault()));
+        plugins.addAll(DynamicPluginLoader.INSTANCE.load(AgentClassLoader.getDefault()));
 
         return plugins;
