+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/util/MethodUtil.java
@@ -16,22 +16,30 @@
  *
  */
 
-package org.apache.skywalking.apm.util;
-
-import org.junit.Assert;
-import org.junit.Test;
+package org.apache.skywalking.apm.agent.core.util;
 
 import java.lang.reflect.Method;
 
 /**
+ * According to the input parameter,
+ * return the OperationName for the span record,
+ * It can determine the unique method
+ *
  * @author zhaoyuguang
  */
 
-public class MethodUtilTest {
+public class MethodUtil {
 
-    @Test
-    public void testClassForName() throws NoSuchMethodException {
-        Assert.assertTrue(MethodUtil.generateOperationName(MethodUtil.class.getMethod("generateOperationName", Method.class))
-                .equals("org.apache.skywalking.apm.util.MethodUtil.generateOperationName(java.lang.reflect.Method)"));
+    public static String generateOperationName(Method method) {
+        StringBuilder operationName = new StringBuilder(method.getDeclaringClass().getName() + "." + method.getName() + "(");
+        Class<?>[] parameterTypes = method.getParameterTypes();
+        for (int i = 0; i < parameterTypes.length; i++) {
+            operationName.append(parameterTypes[i].getName());
+            if (i < (parameterTypes.length - 1)) {
+                operationName.append(",");
+            }
+        }
+        operationName.append(")");
+        return operationName.toString();
     }
 }
